% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itkBSplineScatteredPointsToGrid.R
\name{itkBSplineScatteredPointsToGrid}
\alias{itkBSplineScatteredPointsToGrid}
\title{Generate a regular grid of data from scattered points}
\usage{
itkBSplineScatteredPointsToGrid(x, y, origin = NA, spacing = NA,
  size = NA)
}
\arguments{
\item{x}{matrix of input parameter space values}

\item{y}{matrix of input data space values}

\item{origin}{vector of parameter space starting point}

\item{spacing}{vector of parameter spacing between points in grid}

\item{size}{vector of size of output matrix}
}
\value{
A matrix of regularly space data values
}
\description{
Generate a regular grid of data from scattered points using BSplines
}
\examples{
\dontrun{
t = as.matrix(seq(from=0.0, to=1.0+1e-10, by=0.05))
V0 = 0.25 * cos(t*6.0*3.141)
V1 = 0.25 * sin(t*6.0*3.141)
V2 = 4.00 * t
V = cbind(V0,V1,V2)
VGrid = itkBSplineScatteredPointsToGrid(t,V,origin=0, size=101, spacing=0.01)
}
}
\author{
Duda JT
}

